api/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
flask-cors==4.0.0
gunicorn==21.2.0
scikit-learn==1.3.2